site-admin: notice edit bug fixed

+++ admin/notice-edit.php
@@ -2,6 +2,8 @@
 
     include_once './../config.php';
     include_once 'auth.php';
+
+    $file_error = false;
    
     if(isset($_GET['id'])){
 
@@ -16,28 +18,36 @@
 
     if(isset($_POST['update_btn'])){
 
+
+      if($_FILES['n_file']['size'] < 5*1024*1024){
+
         $n_id = $_POST['n_id'];
         $n_title = $_POST['n_title'];
         $n_description = $_POST['n_description'];
-        $n_file = $_POST['n_file'];
+        $n_file = $_POST['old_file'];
         $n_date = $_POST['n_date'];
 
         if($_FILES['n_file']['size'] > 0){
 
+          $n_file = $_POST['n_file'];
+
           $dir_name = 'uploads/notice/';
           if (!file_exists($dir_name)) { mkdir($dir_name, 0755, true); }
-          $file_name = time() .'.'. pathinfo( $_FILES['$n_file']['name'], PATHINFO_EXTENSION );
-          move_uploaded_file($_FILES['$n_file']['tmp_name'], $dir_name.$file_name);
+
+          $file_name = time() .'.'. pathinfo($_FILES['n_file']['name'], PATHINFO_EXTENSION);
+          move_uploaded_file($_FILES['n_file']['tmp_name'], $dir_name.$file_name);
+
           if(file_exists($_POST['old_file'])){
             unlink($_POST['old_file']);
           }
-          $n_file = $file_name;
+
+          $n_file = $dir_name.$file_name;
 
         }
 
         
 
-        $sql = "UPDATE `notice` SET `title`='$n_title',`description`='$n_description',`date`='$n_date',`link`='$n_file' WHERE id = $n_id";
+        $sql = "UPDATE `notice` SET `title`='$n_title',`description`='$n_description',`date`='$n_date',`file`='$n_file' WHERE id = $n_id";
         $result = mysqli_query($conn, $sql) or die("Query Failed: ". mysqli_error($conn));
 
         if($result){
@@ -46,6 +56,10 @@
         else{
             echo "<script>Update Added Failed </script>";
         }
+      }
+      else{
+        $file_error = true;
+      }
 
     }
 
@@ -85,8 +99,15 @@
                     
                     <hr class="mb-5">
 
+                    <?php if ($file_error): ?>
+                      <div class="alert alert-warning mb-5" role="alert">
+                        <h4>You Have Error!</h4> 
+                        Select a valid file with less than 5MB size.
+                      </div>
+                    <?php endif; ?>
+
                     
-                    <form action="<?php echo $_SERVER['PHP_SELF'].'?id='.$id ?>" method="post" enctype="multipart/form-data">
+                    <form action="<?php echo $_SERVER['PHP_SELF'].'?id='.$n_id ?>" method="post" enctype="multipart/form-data">
                     
                         <div class="form-group">
                             <label>Notice Title</label>
@@ -101,7 +122,11 @@
                         <div class="form-group">
                             <label>Notice File</label>
                             <input name="old_file" type="hidden" value="<?php echo $data['file'] ?>">
-                            <input name="n_file" type="file" class="form-control form-control-lg">
+                            <input name="n_file" type="file" accept=".jpeg,.jpg,.png,.pdf" class="form-control form-control-lg">
+
+                            <?php if(!empty($data['file'])): ?>
+                                <p class="mt-2 mb-4" >Has An Old File: <?php echo $data['file']; ?></p>
+                            <?php endif; ?>
                         </div>
 
                         <div class="form-group">
@@ -110,9 +135,9 @@
                         </div>
                         
                         <input type="hidden" name="n_id" value="<?php echo $data['id'] ?>">
+
                         <input type="submit" value="Save Changes" name="update_btn" class="btn btn-primary">
                     </form>
-
 
 
                   </div>
